changed from int to float

+++ main.py
@@ -13,8 +13,8 @@ def read_root():
 @app.get("/add/{a}/{b}", status_code=200)
 def add(a: str, b: str):
     try:
-        a = int(a)
-        b = int(b)
+        a = float(a)
+        b = float(b)
     except ValueError:
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Both a and b must be valid numbers")
 
@@ -23,8 +23,8 @@ def add(a: str, b: str):
 @app.get("/subtract/{a}/{b}", status_code=200)
 def subtract(a: str, b: str):
     try:
-        a = int(a)
-        b = int(b)
+        a = float(a)
+        b = float(b)
     except ValueError:
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Both a and b must be valid numbers")
 
@@ -33,8 +33,8 @@ def subtract(a: str, b: str):
 @app.get("/multiply/{a}/{b}", status_code=200)
 def multiply(a: str, b: str):
     try:
-        a = int(a)
-        b = int(b)
+        a = float(a)
+        b = float(b)
     except ValueError:
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Both a and b must be valid numbers")
     return {"result": a * b}
@@ -42,8 +42,8 @@ def multiply(a: str, b: str):
 @app.get("/divide/{a}/{b}", status_code=200)
 def divide(a: str, b: str):
     try:
-        a = int(a)
-        b = int(b)
+        a = float(a)
+        b = float(b)
     except ValueError:
         raise HTTPException(status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail="Both a and b must be valid numbers")
     if (b == 0):
